+++ cpp/src/arrow/flight/type_fwd.h
@@ -24,6 +24,7 @@ class Uri;
 namespace flight {
 struct Action;
 struct ActionType;
+struct BasicAuth;
 class ClientAuthHandler;
 class ClientMiddleware;
 class ClientMiddlewareFactory;
@@ -31,6 +32,7 @@ struct Criteria;
 class FlightCallOptions;
 struct FlightClientOptions;
 struct FlightDescriptor;
+struct FlightEndpoint;
 class FlightInfo;
 class FlightListing;
 class FlightMetadataReader;
@@ -41,6 +43,7 @@ class FlightServerOptions;
 class FlightStreamReader;
 class FlightStreamWriter;
 struct Location;
+struct Result;
 class ResultStream;
 struct SchemaResult;
 class ServerCallContext;
